removed soft fork activation & split or merge stake checking
This soft fork activation was added without even an attempt at getting social consensus.  There is no risk of not splitting stakes AFAIK once a network has a number of stakers and far beyond the number of UTXOs it takes to run the network. PIVX even lets you pick the point at which stakes get split.  (todo item?)

On the other hand, since staking on CloakCoin stake rewards paid out on an APR basis of 6% (not a fixed block reward) more competition in staking leads to a more resilient network because bigger UTXOs are getting staked.